Membuat tampilan list

+++ app/src/main/java/com/evanpratama0137/ciciluang/model/Tabungan.kt
@@ -3,7 +3,7 @@ package com.evanpratama0137.ciciluang.model
 data class Tabungan(
     val id: Long,
     val nama: String,
-    val jumlah: Double,
+    val jumlah: Int,
     val deskripsi: String,
     val tanggal: String
 )
